changing property number from int to Long

+++ src/main/java/com/brunobandeira/sistema_de_internacao/entity/Medic.java
@@ -24,7 +24,7 @@ public class Medic {
     private String email;
     @NotNull
     @Min(11)
-    private int number;
+    private Long number;
 
     public UUID getId() {
         return id;
@@ -66,11 +66,11 @@ public class Medic {
         this.email = email;
     }
 
-    public int getNumber() {
+    public Long getNumber() {
         return number;
     }
 
-    public void setNumber(int number) {
+    public void setNumber(Long number) {
         this.number = number;
     }
 }
